remove installation of git
As far as I know it is not necessary to install git to use birdwatcher.
Git is useful for the user for development, but then you can install git also
seperately from anaconda. Only if your preference is to use git within
the conda enviroment you can install it, but that's up to the user.