remove seperate installation of darr and opencv
Darr and opencv are installed directly when installing the birdwatcher package
with git (command "pip install git+https://github.com/gbeckers/birdwatcher@master").
Thus it is unnecessary to install them seperately. I decided to add an 'install requirements' list at
the top of the installation description to inform the user of these dependencies.
As also numpy and matplotlib are installed during this process, I added them to the list as well.